images_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.32
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.8.2
psycopg2-binary==2.9.9
redis==7.1.0
prometheus-fastapi-instrumentator==7.1.0
python-json-logger==2.0.7
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
httpx==0.27.2